% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{dotMatrix}
\alias{dotMatrix}
\title{Performs product of matrices or vectors}
\usage{
dotMatrix(matricesList)
}
\arguments{
\item{matricesList}{a list of numeric matrices with same dimensions}
}
\value{
A matrix in which element (i,j) is the product of all corresponding elements
(i,j) of all matrices in the original list
}
\description{
Given a list of matrices, returns a single matrix in which
    element (i, j) is the product of all corresponding elements (i, j) of
    all matrices in the list.
    Given a list of vectors, returns a single vector in which element i is
    the product of all corresponding elements i of all vectors in the list.
}
