% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageGrid.R
\name{getCounts}
\alias{getCounts}
\title{Get dataframe of counts}
\usage{
# Call getCounts
# getCounts(Grid2d, indexCell)
}
\arguments{
\item{grid}{Grid1d, Grid2d or Grid3d object, made with makeGrid* function}

\item{indexCell}{vector of index made with getCell()}
}
\value{
A dataframe with coordinates and counts
}
\description{
Given a grid object (Grid1d, Grid2d or Grid3d) and a index of
    each cell point in the grid obtained by getCell(), return a dataframe
    with coordinates and counts for each cell.
}
