% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGrid.R
\docType{class}
\name{Grid1d-class}
\alias{Grid1d-class}
\title{Define Grid1d}
\description{
Define Grid1d
}
\section{Slots}{

\describe{
\item{\code{xcell}}{numeric}

\item{\code{xmin}}{numeric}

\item{\code{xmax}}{numeric}
}}

