% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodGrid.R
\name{makeGrid1d}
\alias{makeGrid1d}
\title{Method for Grid1d class}
\usage{
# Create S4 object with the following default values
# makeGrid1d(xmin = -5, xmax = 5, xcell = 100)
}
\arguments{
\item{xcell}{numeric, number of cells}

\item{xmin}{numeric, lower limit}

\item{xmax}{numeric, upper limit}
}
\value{
A uni-dimensional grid
}
\description{
Given the extremes of the range and the number of cells, it
    generates a one-dimensional grid.
}
\examples{
grid1d <- makeGrid1d(xmin = -50, xmax = 50, xcell = 100)
grid1d
> class      : Grid1d
> dimensions : xcell = 100
> range      : xmin = -50, xmax = 50
}
