% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGrid.R
\docType{class}
\name{makeGrid2d}
\alias{makeGrid2d}
\title{Grid2d class}
\value{
A two-dimensional grid
}
\description{
Given the extremes of the range and the number of cells along x
    and y, it generates a two-dimensional grid. Cells are counted from the
    minimum of x and y (bottom-left). You can choose whether to increase
    x or y faster.
}
\section{Slots}{

\describe{
\item{\code{xcell}}{numeric, number of cells along x}

\item{\code{ycell}}{numeric, number of cells along y}

\item{\code{xmin}}{numeric, lower limit along x}

\item{\code{xmax}}{numeric, upper limit along x}

\item{\code{ymin}}{numeric, lower limit along y}

\item{\code{ymax}}{numeric, upper limit along y}

\item{\code{by}}{character, count cells increasing x ("h") or y ("v") faster}
}}

\examples{
grid2d <- makeGrid2d(
            xmin = -50, xmax = 50, xcell = 100,
            ymin = -50, ymax = 50, ycell = 100,
            by = "v"
          )
grid2d

makeGrid2d()

}
