% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGrid.R
\docType{class}
\name{makeGrid3d}
\alias{makeGrid3d}
\title{Grid3d class}
\usage{
# Create S4 object with the following default values
# makeGrid1d(xmin = -5, xmax = 5,
#            ymin = -5, ymax = 5,
#            zmin = -5, zmax = 5,
#            xcell = 10, ycell = 10, zcell = 10)
}
\value{
A three-dimensional grid
}
\description{
Given the extremes of the range and the number of cells along x
    y, and z it generates a three-dimensional grid.
}
\section{Slots}{

\describe{
\item{\code{xcell}}{xcell numeric, number of cells along x}

\item{\code{ycell}}{ycell numeric, number of cells along y}

\item{\code{zcell}}{zcell numeric, number of cells along z}

\item{\code{xmin}}{numeric, lower limit along x}

\item{\code{xmax}}{numeric, upper limit along x}

\item{\code{ymin}}{numeric, lower limit along y}

\item{\code{ymax}}{numeric, upper limit along y}

\item{\code{zmin}}{numeric, lower limit along z}

\item{\code{zmax}}{numeric, upper limit along z}
}}

\examples{
# grid3d <- makeGrid3d(xmin = -50, xmax = 50,
#                      ymin = -50, ymax = 50,
#                      zmin = -50, zmax = 50,
#                      xcell = 100, ycell = 100, zcell = 100)
# grid3d
# class      : Grid3d
# dimensions : xcell = 100, ycell = 100, zcell = 100, ncell = 1e+06
# range      : xmin = -50, xmax = 50
#              ymin = -50, ymax = 50
#              zmin = -50, zmax = 50
}
