% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{pileMatrix}
\alias{pileMatrix}
\title{Trasform matrix into dataframe}
\usage{
pileMatrix(mat, subset = "full")
}
\arguments{
\item{mat}{numeric matrix}

\item{subset}{character indicating which array elements to take. The possible choices are:
"full": the whole matrix (default),
"u": upper triangular matrix without diagonal,
"ud": superior trinagular matrix with diagonal,
"l": lower triangular matrix without diagonal,
"ld": lower triangular matrix with diagonal,
"d": only the diagonal}
}
\value{
A dataframe with three columns: row index (row), col index (col) and matrix element value (value)
}
\description{
Given a matrix, returns a dataframe with three columns, where the
first column is the row index of the original matrix, the second column is the
column index of the original matrix, and the third column is the value of the
original matrix element.
}
\examples{
mat <- matrix(1:9, nrow = 3)

pileMatrix(mat)
#   row col value
# 1   1   1     1
# 2   2   1     2
# 3   3   1     3
# 4   1   2     4
# 5   2   2     5
# 6   3   2     6
# 7   1   3     7
# 8   2   3     8
# 9   3   3     9

pileMatrix(mat, subset = "u")
# row col value
# 1   1   2     4
# 2   1   3     7
# 3   2   3     8

}
