% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin_rowcol.R
\name{rowMax}
\alias{rowMax}
\title{Find max for each row}
\usage{
# Call rowMax
# rowMax(mat)
}
\arguments{
\item{mat}{Matrix object}
}
\value{
A dataframe with values and indices of max for each row
}
\description{
Given a matrix, it returns a dataframe with two column:
    - first column are max values for each row of the original matrix
    - second column are indices of max values
}
