% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmin_rowcol.R
\name{rowMin}
\alias{rowMin}
\title{Find min for each row}
\usage{
rowMin(mat)
}
\arguments{
\item{mat}{Matrix object}
}
\value{
A dataframe with values and indices of min for each row
}
\description{
Given a matrix, it returns a dataframe with two column:
    - first column are min values for each row of the original matrix
    - second column are indices of min values
}
\examples{
mat <- matrix(1:9, nrow = 3)

rowMin(mat)

}
