% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{sumMatrix}
\alias{sumMatrix}
\title{Performs sum of matrices or vectors}
\usage{
sumMatrix(matricesList)
}
\arguments{
\item{matricesList}{a list of numeric matrices with same dimensions}
}
\value{
A matrix in which element (i,j) is the sum of all corresponding elements
(i,j) of all matrices in the original list
}
\description{
Given a list of matrices, returns a single matrix in which
    element (i, j) is the sum of all corresponding elements (i, j) of
    all matrices in the list.
    Given a list of vectors, returns a single vector in which element i is
    the sum of all corresponding elements i of all vectors in the list.
}
\examples{
mat1 <- matrix(1:4, nrow = 2)
mat2 <- matrix(5:8, nrow = 2)
mat3 <- matrix(9:12, nrow = 2)
list_matrix <- list(mat1, mat2, mat3)

sumMatrix(list_matrix)

}
