% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmax}
\alias{blockmax}
\title{Performs block max}
\usage{
blockmax(mat, block)
}
\arguments{
\item{mat}{numeric matrix}

\item{block}{integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A new reduced matrix with max value of each block

#' @examples
mat <- matrix(1:64, nrow = 8, byrow = TRUE)

blockmax(mat, 2) # or blockmax(mat, c(2, 2))
#      [,1] [,2] [,3] [,4]
# [1,]   10   12   14   16
# [2,]   26   28   30   32
# [3,]   42   44   46   48
# [4,]   58   60   62   64
}
\description{
Given a matrix and the size of a block, returns a new matrix
    containing the max value for each block.
}
