% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmean}
\alias{blockmean}
\title{Performs block mean}
\usage{
# Call blockmean
# blockmean(mat, block)
}
\arguments{
\item{mat}{Matrix object}

\item{block}{Integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A new reduced matrix with mean of each block
}
\description{
Given a matrix and the size of a block, returns a new matrix
    containing the average of each block.
}
