% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmean}
\alias{blockmean}
\title{Performs block mean}
\usage{
blockmean(mat, block)
}
\arguments{
\item{mat}{numeric matrix}

\item{block}{integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A new reduced matrix with mean of each block
}
\description{
Given a matrix and the size of a block, returns a new matrix
    containing the average of each block.
}
\examples{
mat <- matrix(1:64, nrow = 8, byrow = TRUE)
mat

blockmean(mat, 2) # or blockmean(mat, c(2, 2))

}
