% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmin}
\alias{blockmin}
\title{Performs block min}
\usage{
blockmin(mat, block)
}
\arguments{
\item{mat}{numeric matrix}

\item{block}{integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A new reduced matrix with min value of each block
}
\description{
Given a matrix and the size of a block, returns a new matrix
    containing the min value for each block.
}
\examples{
mat <- matrix(1:64, nrow = 8, byrow = TRUE)

blockmin(mat, 2) # or blockmin(mat, c(2, 2))
#      [,1] [,2] [,3] [,4]
# [1,]    1    3    5    7
# [2,]   17   19   21   23
# [3,]   33   35   37   39
# [4,]   49   51   53   55

}
