% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{blockmin}
\alias{blockmin}
\title{Performs block min}
\usage{
blockmin(mat, block)
}
\arguments{
\item{mat}{numeric matrix}

\item{block}{integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A new reduced matrix with min value of each block
}
\description{
Given a matrix and the size of a block, returns a new matrix
    containing the min value for each block.
}
