% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{dblocklist}
\alias{dblocklist}
\title{Performs list of diagonal submatrices}
\usage{
# Call dblocklist
# dblockmin(mat, block)
}
\arguments{
\item{mat}{Matrix object}

\item{block}{Integer vector of length 2, containing the size of the block
(rows, columns). If only one integer is passed, the block is square}
}
\value{
A list of diangonal submatrices made by blocks
}
\description{
Given a matrix and the size of a block, returns a list
    containing all diagonal submatrices made by blocks.
}
