% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageGrid.R
\name{getBoxplot}
\alias{getBoxplot}
\title{Get dataframe to boxplot}
\usage{
# Call getBoxplot
# getBoxplot(Grid1d, df_points)
}
\arguments{
\item{grid}{Grid1d object, made with makeGrid1d function}

\item{points}{2d matrix or dataframe of points}
}
\value{
A dataframe with three columns: xbp (x relative to boxplot), x and y
(original points)
}
\description{
Given a Grid1d class object and a 2d dataframe of points
    return original dataframe with supplementary column that group y points
    to boxplot analysis.
}
