% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageGrid.R
\name{getCell}
\alias{getCell}
\title{Get index of a point}
\usage{
# Call getCell
# getCell(Grid2d, points)
}
\arguments{
\item{grid}{Grid1d, Grid2d or Grid3d object, made with makeGrid* function}

\item{points}{vector for 1d Grid, matrix or dataframe of points for 2d 3d Grid}
}
\value{
A vector of index
}
\description{
Given a grid object (Grid1d, Grid2d or Grid3d) and a matrix or
    dataframe of points, return the index of each point in the grid. For a
    1d grid it is counted from left to right; for a 2d grid it is counted
    from top-left corner to bottom-right corner; for a 3d grid for each layer
    it is counted like 2d grid and layers increase from top to bottom
}
