% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices_dfs.R
\name{getTriang}
\alias{getTriang}
\title{Get triangular items of a grid}
\usage{
getTriang(ncell, part = "upper", mirror = FALSE, diag = TRUE, by = "h")
}
\arguments{
\item{ncell}{integer indicating number of cells in grid}

\item{part}{character indicating which part of grid to take. The possible choices
are "upper" and "lower"}

\item{mirror}{boolean indicating whether to take the main direction (FALSE) or
the mirrored one (TRUE). Default FALSE}

\item{diag}{boolean indicating wheater include (TRUE) diagonal or not (FALSE).
Default TRUE}

\item{by}{character indicating whether grid indices grow faster along x ("h") or y ("v")}
}
\value{
A vector of indices
}
\description{
Given the number of elements of a square grid, it returns the
indices of the upper or lower triangular part, with or without diagonal, both
in the main direction and in the mirrored direction. The item count starts
from the bottom left and proceeds by incrementing x ("h") or y ("v") faster.
}
\examples{
df_grid <- expand.grid(x = 1:4, y = 1:4)
df_grid$index <- 1:16
df_grid

# corresponding grid
# 13 14 15 16
#  9 10 11 12
#  5  6  7  8
#  1  2  3  4

getTriang(nrow(df_grid), part = "upper", mirror = TRUE, diag = TRUE, by = "h")

}
