% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodGrid.R
\name{makeGrid2d}
\alias{makeGrid2d}
\title{Method for Grid2d class}
\usage{
# Create S4 object with the following default values
# makeGrid1d(xmin = -5, xmax = 5,
             ymin = -5, ymax = 5,
             xcell = 100, ycell = 100)
}
\arguments{
\item{xcell}{numeric, number of cells along x}

\item{ycell}{numeric, number of cells along y}

\item{xmin}{numeric, lower limit along x}

\item{xmax}{numeric, upper limit along x}

\item{ymin}{numeric, lower limit along y}

\item{ymax}{numeric, upper limit along y}
}
\value{
A two-dimensional grid
}
\description{
Given the extremes of the range and the number of cells along x
    and y, it generates a two-dimensional grid.
}
\examples{
grid2d <- makeGrid2d(xmin = -50, xmax = 50,
                     ymin = -50, ymax = 50,
                     xcell = 100, ycell = 100)
grid2d
> class      : Grid2d
> dimensions : xcell = 100, ycell = 100, ncell = 10000
> range      : xmin = -50, xmax = 50
               ymin = -50, ymax = 50
}
