% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classGrid.R
\docType{class}
\name{makeGrid3d}
\alias{makeGrid3d}
\title{Grid3d class}
\value{
A three-dimensional grid
}
\description{
Given the extremes of the range and the number of cells along x
    y and z, it generates a three-dimensional grid. Cells are counted from the
    minimum of x, y and z. You can choose whether to increase x or z faster.
}
\section{Slots}{

\describe{
\item{\code{xcell}}{numeric, number of cells along x}

\item{\code{ycell}}{numeric, number of cells along y}

\item{\code{zcell}}{numeric, number of cells along z}

\item{\code{xmin}}{numeric, lower limit along x}

\item{\code{xmax}}{numeric, upper limit along x}

\item{\code{ymin}}{numeric, lower limit along y}

\item{\code{ymax}}{numeric, upper limit along y}

\item{\code{zmin}}{numeric, lower limit along z}

\item{\code{zmax}}{numeric, upper limit along z}

\item{\code{by}}{character, count cells increasing x ("h") or z ("v") faster}
}}

\examples{
grid3d <- makeGrid3d(xmin = -50, xmax = 50,
                     ymin = -50, ymax = 50,
                     zmin = -50, zmax = 50,
                     xcell = 4, ycell = 5, zcell = 6,
                     by = "v")

grid3d
# class      : Grid3d
# dimensions : xcell = 4, ycell = 5, zcell = 6, ncell = 120
# range      : xmin = -50, xmax = 50
#              ymin = -50, ymax = 50
#              zmin = -50, zmax = 50
# by         : v, count starts from xmin, ymin, zmin;
#              z increase faster, x slower"


makeGrid3d()
# class      : Grid3d
# dimensions : xcell = 5, ycell = 5, zcell = 5, ncell = 125
# range      : xmin = -5, xmax = 5
#              ymin = -5, ymax = 5
#              zmin = -5, zmax = 5
# by         : h, count starts from xmin, ymin, zmin;
#              x increase faster, z slower"
}
