% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodGrid.R
\name{makeGrid3d}
\alias{makeGrid3d}
\title{Method for Grid3d class}
\usage{
# Create S4 object with the following default values
# makeGrid1d(xmin = -5, xmax = 5,
             ymin = -5, ymax = 5,
             zmin = -5, zmax = 5,
             xcell = 100, ycell = 100, zcell = 100)
}
\arguments{
\item{xcell}{numeric, number of cells along x}

\item{ycell}{numeric, number of cells along y}

\item{zcell}{numeric, number of cells along z}

\item{xmin}{numeric, lower limit along x}

\item{xmax}{numeric, upper limit along x}

\item{ymin}{numeric, lower limit along y}

\item{ymax}{numeric, upper limit along y}

\item{zmin}{numeric, lower limit along z}

\item{zmax}{numeric, upper limit along z}
}
\value{
A three-dimensional grid
}
\description{
Given the extremes of the range and the number of cells along x
    y, and z it generates a three-dimensional grid.
}
\examples{
grid3d <- makeGrid3d(xmin = -50, xmax = 50,
                     ymin = -50, ymax = 50,
                     zmin = -50, zmax = 50,
                     xcell = 100, ycell = 100, zcell = 100)
grid3d
> class      : Grid3d
> dimensions : xcell = 100, ycell = 100, zcell = 100, ncell = 1e6
> range      : xmin = -50, xmax = 50
               ymin = -50, ymax = 50
               zmin = -50, zmax = 50
}
