% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableToLatex.R
\name{tableToLatex}
\alias{tableToLatex}
\title{Convert R table in LaTeX table}
\usage{
# Call tableToLatex
# tableToLatex(object)
}
\arguments{
\item{object}{A numeric matrix, data.frame, tibble or data.table}

\item{digits}{Number of decimals. Default = 3}

\item{file}{If TRUE write file with object name in work directory. Default FALSE}

\item{double_space}{If TRUE add empty columns to increase space between columns}
}
\value{
Text on terminal or on a file to copy in LaTeX environment
}
\description{
Given a numeric matrix, a data.frame, a tibble or a data.table
    returns a basic LaTeX table write with table and tabular packages.
    Rownames and colnames are highlighted with bold.
}
